pyproject.toml
[project]
name = "scientific-file-format-data-analyzer"
version = "0.1.0"
description = "Analyzes and visualizes file format distributions across scientific data repositories including Data.gov, ESS-DIVE, HuggingFace, and IEEE Dataport."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.35.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
